+++ lib/socket.ts
@@ -1,15 +1,11 @@
-import * as Bull from "bull";
 import { RedisOptions } from "ioredis";
 import { pick } from "lodash";
 import * as url from "url";
-import {
-  getCache,
-  updateQueuesCache,
-  getRedisInfo,
-  ping,
-  queueKey,
-} from "./queues-cache";
+import { getCache, updateQueuesCache, queueKey } from "./queues-cache";
 import { WebSocketClient } from "./ws-autoreconnect";
+import { execRedisCommand, getRedisInfo, ping } from "./queue-factory";
+import { getQueueType } from "./utils";
+import { Integration } from "./interfaces/integration";
 
 const { version } = require(`${__dirname}/../package.json`);
 
@@ -29,9 +25,15 @@ export const Socket = (
   server: string,
   token: string,
   connection: Connection,
-  team?: string,
-  nodes?: string[]
+  opts: {
+    team?: string;
+    nodes?: string[];
+    integrations?: {
+      [key: string]: Integration;
+    };
+  } = {}
 ) => {
+  const { team, nodes, integrations } = opts;
   const ws = new WebSocketClient();
   const redisOpts = redisOptsFromConnection(connection);
 
@@ -43,9 +45,9 @@ export const Socket = (
   });
 
   console.log(
-    chalk.yellow("WebSocket:") +
-      chalk.blueBright(" opening connection to ") +
-      chalk.gray("Taskforce.sh")
+    `${chalk.yellow("WebSocket:")} ${chalk.blueBright(
+      "opening connection to"
+    )} ${chalk.gray("Taskforce.sh")} (${chalk.blueBright(server)})`
   );
 
   ws.onopen = function open() {
@@ -72,219 +74,91 @@ export const Socket = (
       `${chalk.yellow("WebSocket:")} ${chalk.blueBright("received")} %s`,
       input
     );
-    if (input === "authorized") {
-      console.log(
-        chalk.yellow("WebSocket: ") +
-          chalk.green("Succesfully authorized to taskforce.sh service")
-      );
 
-      //
-      // Send this connection.
-      //
-      const queues = await updateQueuesCache(redisOpts, nodes);
-      console.log(
-        `${chalk.yellow("WebSocket: ")} ${chalk.green(
-          "sending connection: "
-        )} ${chalk.blueBright(name)} ${
-          team ? chalk.green(" for team ") + chalk.blueBright(team) : ""
-        }`
-      );
-      ws.send(
-        JSON.stringify({
-          res: "connection",
-          cmd: "update",
-          queues,
-          connection: name,
-          team,
-          version,
-        })
-      );
-    } else {
-      const msg = JSON.parse(input);
-
-      if (!msg.data) {
-        console.error(
-          chalk.red("WebSocket:") + chalk.blueBright(" missing message data "),
-          msg
+    try {
+      if (input === "authorized") {
+        console.log(
+          chalk.yellow("WebSocket: ") +
+            chalk.green("Succesfully authorized to taskforce.sh service")
         );
-        return;
-      }
 
-      const { res, queueName, queuePrefix } = msg.data;
+        //
+        // Send this connection.
+        //
+        const queues = await updateQueuesCache(redisOpts, {
+          nodes,
+          integrations,
+        });
+        console.log(
+          `${chalk.yellow("WebSocket: ")} ${chalk.green(
+            "sending connection: "
+          )} ${chalk.blueBright(name)} ${
+            team ? chalk.green(" for team ") + chalk.blueBright(team) : ""
+          }`
+        );
+        ws.send(
+          JSON.stringify({
+            res: "connection",
+            cmd: "update",
+            queues,
+            connection: name,
+            team,
+            version,
+          })
+        );
+      } else {
+        const msg = JSON.parse(input);
 
-      switch (res) {
-        case "connections":
-          respondConnectionCommand(connection, msg);
-          break;
-        case "queues":
-        case "jobs":
-          const cache = getCache();
-          if (!cache) {
-            await updateQueuesCache(redisOpts, nodes);
-          }
-          const queue =
-            cache[queueKey({ name: queueName, prefix: queuePrefix || "bull" })];
+        if (!msg.data) {
+          console.error(
+            chalk.red("WebSocket:") +
+              chalk.blueBright(" missing message data "),
+            msg
+          );
+          return;
+        }
 
-          if (!queue) {
-            ws.send(
-              JSON.stringify({
-                id: msg.id,
-                err: "Queue not found",
-              })
-            );
-          } else {
-            switch (res) {
-              case "queues":
-                respondQueueCommand(queue, msg);
-                break;
-              case "jobs":
-                respondJobCommand(queue, msg);
-                break;
+        const { res, queueName, queuePrefix } = msg.data;
+
+        switch (res) {
+          case "connections":
+            await respondConnectionCommand(connection, msg);
+            break;
+          case "queues":
+          case "jobs":
+            const cache = getCache();
+            if (!cache) {
+              await updateQueuesCache(redisOpts, opts);
             }
-          }
-          break;
+            const { queue, responders } =
+              cache[
+                queueKey({ name: queueName, prefix: queuePrefix || "bull" })
+              ];
+
+            if (!queue) {
+              ws.send(
+                JSON.stringify({
+                  id: msg.id,
+                  err: "Queue not found",
+                })
+              );
+            } else {
+              switch (res) {
+                case "queues":
+                  await responders.respondQueueCommand(ws, queue, msg);
+                  break;
+                case "jobs":
+                  await responders.respondJobCommand(ws, queue, msg);
+                  break;
+              }
+            }
+            break;
+        }
       }
+    } catch (err) {
+      console.error(err);
     }
   };
-
-  function paginate(
-    queue: Bull.Queue,
-    messageId: string,
-    start: number,
-    end: number,
-    method: string,
-    opts?: {
-      excludeData: boolean;
-    }
-  ) {
-    start = start || 0;
-    end = end || -1;
-    return (<any>queue)
-      [method](start, end, opts)
-      .then(function (jobs: Bull.Job[]) {
-        respond(messageId, jobs);
-      });
-  }
-
-  async function respondJobCommand(queue: Bull.Queue, msg: any) {
-    const data = msg.data;
-    const job = await queue.getJob(data.jobId);
-
-    switch (data.cmd) {
-      case "retry":
-        await job.retry();
-        break;
-      case "promote":
-        await job.promote();
-        break;
-      case "remove":
-        await job.remove();
-        break;
-      case "discard":
-        await job.discard();
-        break;
-      case "moveToFailed":
-        await job.moveToFailed({ message: "Failed manually" });
-        break;
-      case "update":
-        await job.update(data.data);
-      default:
-        console.error(
-          `Missing command ${data.cmd}. Too old version of taskforce-connector?`
-        );
-    }
-    respond(msg.id);
-  }
-
-  async function respondQueueCommand(queue: Bull.Queue, msg: any) {
-    const data = msg.data;
-    switch (data.cmd) {
-      case "getJob":
-        const job = await queue.getJob(data.jobId);
-        respond(msg.id, job);
-        break;
-      case "getJobCounts":
-        const jobCounts = await queue.getJobCounts();
-        respond(msg.id, jobCounts);
-        break;
-      case "getMetrics":
-        const metrics = await (<any>queue).getMetrics(
-          data.type,
-          data.start,
-          data.end
-        );
-        respond(msg.id, metrics);
-        break;
-      case "getWaiting":
-      case "getActive":
-      case "getDelayed":
-      case "getCompleted":
-      case "getFailed":
-      case "getRepeatableJobs":
-      case "getWorkers":
-        paginate(queue, msg.id, data.start, data.end, data.cmd, data.opts);
-        break;
-
-      case "getJobLogs":
-        const logs = await queue.getJobLogs(data.jobId, data.start, data.end);
-        respond(msg.id, logs);
-
-      case "getWaitingCount":
-      case "getActiveCount":
-      case "getDelayedCount":
-      case "getCompletedCount":
-      case "getFailedCount":
-      case "getRepeatableCount":
-      case "getWorkersCount":
-        const count = await (<any>queue)[data.cmd]();
-        respond(msg.id, count);
-        break;
-      case "removeRepeatableByKey":
-        await queue.removeRepeatableByKey(data.key);
-        respond(msg.id);
-        break;
-      case "add":
-        await queue.add(...(data.args as [string, object, object]));
-        respond(msg.id);
-        break;
-      case "empty":
-        await queue.empty();
-        respond(msg.id);
-        break;
-      case "pause":
-        await queue.pause();
-        respond(msg.id);
-        break;
-      case "resume":
-        await queue.resume();
-        respond(msg.id);
-        break;
-      case "isPaused":
-        const isPaused = await queue.isPaused();
-        respond(msg.id, isPaused);
-        break;
-      case "obliterate":
-        await queue.obliterate();
-        respond(msg.id);
-        break;
-      case "clean":
-        await queue.clean(data.grace, data.status, data.limit);
-        respond(msg.id);
-        break;
-      case "retryJobs":
-        await (<any>queue).retryJobs({
-          status: data.status,
-          count: data.count,
-        });
-        respond(msg.id);
-        break;
-      default:
-        console.error(
-          `Missing command ${data.cmd}. Too old version of taskforce-connector?`
-        );
-        respond(msg.id, null);
-    }
-  }
 
   async function respondConnectionCommand(connection: Connection, msg: any) {
     const data = msg.data;
@@ -295,7 +169,7 @@ export const Socket = (
         break;
       case "getConnection":
         {
-          const queues = await updateQueuesCache(redisOpts, nodes);
+          const queues = await updateQueuesCache(redisOpts, opts);
 
           console.log(
             `${chalk.yellow("WebSocket: ")} ${chalk.green(
@@ -315,7 +189,7 @@ export const Socket = (
         break;
       case "getQueues":
         {
-          const queues = await updateQueuesCache(redisOpts, nodes);
+          const queues = await updateQueuesCache(redisOpts, opts);
 
           console.log(
             `${chalk.yellow("WebSocket:")} ${chalk.blueBright(
@@ -330,6 +204,15 @@ export const Socket = (
       case "getInfo":
         const info = await getRedisInfo(redisOpts, nodes);
         respond(msg.id, info);
+        break;
+
+      case "getQueueType":
+        const queueType = await execRedisCommand(
+          redisOpts,
+          (client) => getQueueType(data.name, data.prefix, client),
+          nodes
+        );
+        respond(msg.id, { queueType });
         break;
     }
   }
@@ -346,10 +229,12 @@ export const Socket = (
 function redisOptsFromConnection(connection: Connection): RedisOptions {
   let opts: RedisOptions = {
     ...pick(connection, [
-      "port",
       "host",
-      "family",
+      "port",
+      "username",
       "password",
+      "family",
+      "sentinelPassword",
       "db",
       "tls",
       "sentinels",
@@ -380,6 +265,8 @@ function redisOptsFromUrl(urlString: string) {
       ? parseInt(redisUrl.pathname.split("/")[1])
       : 0;
     if (redisUrl.auth) {
+      const username = redisUrl.auth.split(":")[0];
+      redisOpts.username = username ? username : undefined;
       redisOpts.password = redisUrl.auth.split(":")[1];
     }
   } catch (e) {